Fix Test report members

+++ NGitLab/Models/TestReport.cs
@@ -6,7 +6,7 @@ namespace NGitLab.Models;
 public class TestReport
 {
     [JsonPropertyName("total_time")]
-    public int TotalTime { get; set; }
+    public double TotalTime { get; set; }
 
     [JsonPropertyName("total_count")]
     public int TotalCount { get; set; }
